결제 기능 수정(roomNo, residNo/int-> Long으로 수정)

+++ src/main/java/com/springbootfinal/app/domain/room/Reservation.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Reservation {
-    private int roomNo;
+    private Long roomNo;
     private Long userNo;
     private LocalDate checkinDate;
     private LocalDate checkoutDate;
